CONC-312: Implementation of caching_sha2_plugin
This implementation will work with OpenSSL and Schannel only, but not with GnuTLS.
Neither GnuTLS nor libnettle provide the rsa encryption methods for OAEP padding.

Added new option for mysql_optionsv: MYSQL_SERVER_PUBLIC_KEY which spedifys the name
of the file which contains the public rsa key of the database server. The format of the
file must be PEM format. The corresponding key name for the configuration files is "server_public_key".