Fix GCC 8 -Wstringop-truncation
If strncpy() truncates the string due to insufficient destination
buffer, it will not NUL-terminate the destination string.
10e4d47b0cc6003dc87285261b14f87d87ac01bd ensured that the
destination string will always be NUL-terminated, by writing a NUL to
the end of the buffer.

But, it failed to silence a -Wstringop-truncation warning.
This fix (which should be non-functional) actually does that.