MDEV-27304 SHOW ... result columns are right-aligned
with --version=xxx the server doesn't send 5.5.5- version prefix,
so C/C doesn't detect the server as MariaDB (look for "is_maria"),
but it still sends MARIADB_CLIENT_EXTENDED_METADATA to the server.
This causes the server to send extended metadata, but C/C doesn't expect
it (is_maria is false) and parses the packet incorrectly.

As a fix, backport from 3.2 the code that makes C/C to announce
only those client capabilities that are supported by the server.