Fix for CONC-612:
When connecting via socket (non unix-socket) the connection can't
be established if the wait function (usleep) returns an error and
modifies errno. This is the case on MacOSX, where usleep() returns
-1 and sets errno to ETIMED_OUT. This patch saves errno from connect()
call and restores it after usleep.

This bug was introduced with PR204 (commit ef761c1f1393b676208509b7786a6eb3436f5f90)