CONC-774 fixup: Always emit BTI J
In commit 51e3519d6676a548c03e9e89220b22397ec470fe an incorrect
numeric encoding was added that is for the instruction BTI, not BTI J.

According to https://www.spinics.net/lists/arm-kernel/msg1041915.html
the two most significant bits of the least significant octet of the
instruction word for BTI are [J][C] flags.

Because a newer version of GCC could emit code for an older assembler,
we had better always use the numeric encoding on GCC. In clang and
derivatives, the assembler is built-in.